+++ gluecodium/src/test/resources/smoke/durations/output/dart/ffi/ffi_smoke_DurationInterface.cpp
@@ -1,0 +1,119 @@
+#include "ffi_smoke_DurationInterface.h"
+#include "ConversionBase.h"
+#include "InstanceCache.h"
+#include "FinalizerData.h"
+#include "CallbacksQueue.h"
+#include "IsolateContext.h"
+#include "ProxyCache.h"
+#include "gluecodium/DurationHash.h"
+#include "gluecodium/TypeRepository.h"
+#include "smoke/DurationInterface.h"
+#include <chrono>
+#include <memory>
+#include <string>
+#include <memory>
+#include <new>
+class smoke_DurationInterface_Proxy : public smoke::DurationInterface {
+public:
+    smoke_DurationInterface_Proxy(uint64_t token, int32_t isolate_id, Dart_Handle dart_handle, FfiOpaqueHandle f0)
+        : token(token), isolate_id(isolate_id), dart_persistent_handle(Dart_NewPersistentHandle_DL(dart_handle)), f0(f0) {
+        library_cache_dart_handle_by_raw_pointer(this, isolate_id, dart_handle);
+    }
+    ~smoke_DurationInterface_Proxy() {
+        gluecodium::ffi::remove_cached_proxy(token, isolate_id, "smoke_DurationInterface");
+        auto raw_pointer_local = this;
+        auto isolate_id_local = isolate_id;
+        auto dart_persistent_handle_local = dart_persistent_handle;
+        auto deleter = [raw_pointer_local, isolate_id_local, dart_persistent_handle_local]() {
+            library_uncache_dart_handle_by_raw_pointer(raw_pointer_local, isolate_id_local);
+            Dart_DeletePersistentHandle_DL(dart_persistent_handle_local);
+        };
+        if (gluecodium::ffi::IsolateContext::is_current(isolate_id)) {
+            deleter();
+        } else {
+            gluecodium::ffi::cbqm.enqueueCallback(isolate_id, deleter);
+        }
+    }
+    smoke_DurationInterface_Proxy(const smoke_DurationInterface_Proxy&) = delete;
+    smoke_DurationInterface_Proxy& operator=(const smoke_DurationInterface_Proxy&) = delete;
+    std::string
+    duration_function(const std::chrono::seconds input) override {
+        FfiOpaqueHandle _result_handle;
+        dispatch([&]() { (*reinterpret_cast<bool (*)(Dart_Handle, uint64_t, FfiOpaqueHandle*)>(f0))(Dart_HandleFromPersistent_DL(dart_persistent_handle),
+            gluecodium::ffi::Conversion<std::chrono::seconds>::toFfi(input),
+            &_result_handle
+        ); });
+        auto _result = gluecodium::ffi::Conversion<std::string>::toCpp(_result_handle);
+        delete reinterpret_cast<std::string*>(_result_handle);
+        return _result;
+    }
+private:
+    const uint64_t token;
+    const int32_t isolate_id;
+    const Dart_PersistentHandle dart_persistent_handle;
+    const FfiOpaqueHandle f0;
+    inline void dispatch(std::function<void()>&& callback) const
+    {
+        gluecodium::ffi::IsolateContext::is_current(isolate_id)
+            ? callback()
+            : gluecodium::ffi::cbqm.enqueueCallback(isolate_id, std::move(callback)).wait();
+    }
+};
+#ifdef __cplusplus
+extern "C" {
+#endif
+FfiOpaqueHandle
+library_smoke_DurationInterface_durationFunction__Duration(FfiOpaqueHandle _self, int32_t _isolate_id, uint64_t input) {
+    gluecodium::ffi::IsolateContext _isolate_context(_isolate_id);
+    return gluecodium::ffi::Conversion<std::string>::toFfi(
+        (*gluecodium::ffi::Conversion<std::shared_ptr<smoke::DurationInterface>>::toCpp(_self)).duration_function(
+            gluecodium::ffi::Conversion<std::chrono::seconds>::toCpp(input)
+        )
+    );
+}
+// "Private" finalizer, not exposed to be callable from Dart.
+void
+library_smoke_DurationInterface_finalizer(FfiOpaqueHandle handle, int32_t isolate_id) {
+    auto ptr_ptr = reinterpret_cast<std::shared_ptr<smoke::DurationInterface>*>(handle);
+    library_uncache_dart_handle_by_raw_pointer(ptr_ptr->get(), isolate_id);
+    library_smoke_DurationInterface_release_handle(handle);
+}
+void
+library_smoke_DurationInterface_register_finalizer(FfiOpaqueHandle ffi_handle, int32_t isolate_id, Dart_Handle dart_handle) {
+    FinalizerData* data = new (std::nothrow) FinalizerData{ffi_handle, isolate_id, &library_smoke_DurationInterface_finalizer};
+    Dart_NewFinalizableHandle_DL(dart_handle, data, sizeof data, &library_execute_finalizer);
+}
+FfiOpaqueHandle
+library_smoke_DurationInterface_copy_handle(FfiOpaqueHandle handle) {
+    return reinterpret_cast<FfiOpaqueHandle>(
+        new (std::nothrow) std::shared_ptr<smoke::DurationInterface>(
+            *reinterpret_cast<std::shared_ptr<smoke::DurationInterface>*>(handle)
+        )
+    );
+}
+void
+library_smoke_DurationInterface_release_handle(FfiOpaqueHandle handle) {
+    delete reinterpret_cast<std::shared_ptr<smoke::DurationInterface>*>(handle);
+}
+FfiOpaqueHandle
+library_smoke_DurationInterface_create_proxy(uint64_t token, int32_t isolate_id, Dart_Handle dart_handle, FfiOpaqueHandle f0) {
+    auto cached_proxy = gluecodium::ffi::get_cached_proxy<smoke_DurationInterface_Proxy>(token, isolate_id, "smoke_DurationInterface");
+    std::shared_ptr<smoke_DurationInterface_Proxy>* proxy_ptr;
+    if (cached_proxy) {
+        proxy_ptr = new (std::nothrow) std::shared_ptr<smoke_DurationInterface_Proxy>(cached_proxy);
+    } else {
+        proxy_ptr = new (std::nothrow) std::shared_ptr<smoke_DurationInterface_Proxy>(
+            new (std::nothrow) smoke_DurationInterface_Proxy(token, isolate_id, dart_handle, f0)
+        );
+        gluecodium::ffi::cache_proxy(token, isolate_id, "smoke_DurationInterface", *proxy_ptr);
+    }
+    return reinterpret_cast<FfiOpaqueHandle>(proxy_ptr);
+}
+FfiOpaqueHandle
+library_smoke_DurationInterface_get_type_id(FfiOpaqueHandle handle) {
+    const auto& type_id = ::gluecodium::get_type_repository().get_id(reinterpret_cast<std::shared_ptr<smoke::DurationInterface>*>(handle)->get());
+    return reinterpret_cast<FfiOpaqueHandle>(new (std::nothrow) std::string(type_id));
+}
+#ifdef __cplusplus
+}
+#endif